backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.22
deepface==0.0.93
opencv-python-headless==4.9.0.80
spotipy==2.25.1
python-dotenv==1.0.1
numpy==1.26.4
Pillow==12.1.1
tf-keras==2.16.0
google-generativeai==0.8.3
youtube-search-python==1.6.6
httpx>=0.23.0,<0.24.0
